#171 Create test database connections
